Add Mistral and refine LLaMa 2 README
Add a newer complete installation that has Mistral support as well as the
LLaMa2. Keep in mind LLaMa2 now has bugs that were no there in the Sept
edition of text-generation-webui. Bugs mostly related to GPTQ and GPU Memory bloat
are controllable with --disable_exllama --disable_exllamav2 --loader autogptq.
This will force the app to revert back to the pre-Sept GPTQ quantized model loader
which is still in the code-base but is non-default.